more

+++ old_functions.cpp
@@ -44,8 +44,8 @@ void Fiborand(int fib_var[], int max_var, int fractar)
 
 void Dark_Maze_PRNG(int square_matrix[][201], int max_level, int size_level, int clear_radius,
 bool half_gone, bool pillars_exist, bool wall_exist, bool zero_wall, bool wall_concrete, bool exit_exist, bool dark_exist,
-int fib_val[], int max_val, int fractal, double wall_frac, double dark_frac, double exit_frac,
-double candy_frac, double red_candy_frac, double yellow_candy_frac, double green_candy_frac, double blue_candy_frac)
+int fib_val[], int max_val, int fractal, float wall_frac, float dark_frac, float exit_frac,
+float candy_frac, float red_candy_frac, float yellow_candy_frac, float green_candy_frac, float blue_candy_frac)
 {
   for (int a_sub = -size_level; a_sub <= size_level; a_sub++)
   {
@@ -321,7 +321,7 @@ void Color_Picker(int cube_identityy, int colorz[], int multikolor[])
 }
 
 sf::RectangleShape Square_Draw(sf::RectangleShape squaraa, int colourz[], int tranzp,
-                               double poz_x, double poz_y, double squarral_x, double squarral_y)
+                               float poz_x, float poz_y, float squarral_x, float squarral_y)
 {
   squaraa.setPosition(poz_x - squarral_x, poz_y - squarral_y);
   squaraa.setFillColor(sf::Color(colourz[0], colourz[1], colourz[2], tranzp));
